Build 1457:
(1) Configurations are now listed in case-INsensitive alphabetic order and are no longer surrounded by single-quote marks on the drop-down menu.

(2) You can now 'add' configurations in ~/Library/...Configurations to deployed configurations. Thanks go to Michael Williams, who both thought this up and programmed it. This is NOT DONE UNLESS a forced preference named 'useLibraryConfigurationsWithDeployedOnes' (boolean) is set TRUE.

Currently, the only indication that a configuration is from Deploy is that on the Details… window the window title shows '(Deployed)' after the configuration display name. This is done only when in such a mixed-configuration environment. Connections in ~/Library/...Configurations are ignored if they have the same display name as configurations in Deploy. That is subject to change; we may warn with a dialog or put something in the Console log.

Currently, the contents of Deploy and ~/Library/...Configurations are merged into a single, alphabetized list. That is subject to change; we may list the Deploy configurations first, followed by the configurations from ~/Library/...Configurations.

git-svn-id: http://tunnelblick.googlecode.com/svn/trunk@457 ecaf4b9d-c145-0410-a236-1b700048ef5c