Build 1991:
(1) Fixes bug that repeatedly recreates the 'Launch Tunnelblick' link in the configurations folder.

(2) Works around the following OpenVPN bug: the OpenVPN process ignores the first SIGTERM (via kill or management interface) and to delays termination for tens of seconds when in the 'RESOLVE' state.

Works around this by warning the user that this is happening, then repeatedly sending SIGTERM and, after a timeout period (default is 180 seconds), considering the connection closed even if OpenVPN doesn't acknowledge the closing.

Note: Tunnelblick _never_ sends a SIGKILL to terminate an OpenVPN process forcibly; only SIGTERM is used.

(3) Localization update.

git-svn-id: http://tunnelblick.googlecode.com/svn/trunk@991 ecaf4b9d-c145-0410-a236-1b700048ef5c