Build 2206: Integrates 3.1 branch build 2190.2205 fixes into the trunk:
(1) Fixes problem with left navigation sometimes not being displayed when Details... window was open and had tabs last time it was displayed but now doesn't have tabs (because one or more configurations have been added, for example).

(2) Fixes problem (present since initial version of Tunnelblick) with icon not displaying correct status. Now shows closed tunnel if all configurations that the user should expect to be closed are in fact closed and shows open tunnel if all configurations that the user should expect to be open are in fact open; otherwise the animation is shown. Previously displayed the status of the last tunnel which had a change, which can be misleading when more than one tunnel is or should be open at a time.

(3) Fixes problem in error message referring to a known problem in OpenVPN 2.1 -- the error message incorrectly referred to OpenVPN 2.2.

git-svn-id: http://tunnelblick.googlecode.com/svn/trunk@1206 ecaf4b9d-c145-0410-a236-1b700048ef5c